refactor: extract Htlc component to standalone
