+++ renderer/components/Activity/PaymentModal/index.js
@@ -1,2 +1,3 @@
 export PaymentModal from './PaymentModal'
 export Route from './Route'
+export Htlc from './Htlc'
